Fix annoying compiler warning
The compiler gives us a warning that we're comparing between different
signedness in a range check. Add a static_cast to remove it

+++ src/Token/TokenParser.h
@@ -168,7 +168,7 @@ public:
         }
 
         // Check range
-        if (t_pos >= size())
+        if (t_pos >= static_cast<long long int>(size()))
             throw std::range_error("Read pointer >= " + std::to_string(m_tokens.size()) + ".");
         
         return m_tokens.at(t_pos);
